pyproject.toml
[tool.poetry]
name = "tranotra-leads"
version = "1.0.0"
description = "Automated customer discovery and outreach pipeline for Tranotra"
authors = ["Ailong <ailong@example.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "tranotra", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^3.0.0"
SQLAlchemy = "2.0.0"
python-dotenv = "1.0.0"
google-generativeai = "0.3.0"
click = "^8.0"
python-json-logger = "2.0.0"
Werkzeug = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.0"
pytest-cov = "4.1.0"
pytest-mock = "3.11.1"
black = "23.0.0"
isort = "5.0.0"
mypy = "1.0.0"

[tool.poetry.scripts]
tranotra = "tranotra.cli:main"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_mode = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
addopts = "-v --cov=src/tranotra --cov-report=term-missing --cov-fail-under=80"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
